Do not rely on flask.g for read_preference context manager (#5)
Newer versions of flask have flask.g be bound to the application-context. This
means that the read preference would leak in case of processes that create one
application context but use multiple request contexts.

This works with greenlets as well, since gevent.monkey_patch makes
threading.local work with greenlets.